+++ CHANGELOG.md
@@ -4,8 +4,12 @@
 
 ### New Features
 
-* Added support for directly plotting Road to Riches, Exobiology (Expressway to Exomastery), Ammonia World, Earth-like World, and Rocky/High Metal Content World routes
-* Added a Fleet Carrier route plotter
+* Added support for directly plotting:
+  * Road to Riches
+  * Exobiology (Expressway to Exomastery)
+  * Trade Routes
+  * Tourist Routes
+  * Fleet Carrier Routes
 * Added via-point hops to the Neutron plotter, and migrated the Tourist plotter's stop list to the same +/- row UI
 
 ### Changes